requirements/base.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.27.1  # https://github.com/tfranzel/drf-spectacular
djangorestframework-simplejwt==5.3.1  # https://github.com/jpadilla/django-rest-framework-simplejwt

psycopg[c]==3.1.9  # https://github.com/psycopg/psycopg
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
